requirements.txt
agno
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
chromadb
sentence-transformers
pydantic
python-dotenv
streamlit
requests
fastapi
uvicorn
pypdf
unstructured
markdown
tqdm
langchain-chroma
groq

